[mod] use `#[repr(u8)]` for `Finger`&`Hand` enums

+++ layout_evaluation/src/metrics/bigram_metrics/movement_pattern_same_row.rs
@@ -75,7 +75,7 @@ impl BigramMetric for MovementPatternSameRow {
 
         // no roll on lateral finger movements
         if self.exclude_lateral_finger_movement
-            && k1.key.finger.distance(&k2.key.finger) < (pos1.0 as i8 - pos2.0 as i8).abs() as usize
+            && k1.key.finger.distance(&k2.key.finger) < (pos1.0 as i8 - pos2.0 as i8).abs() as u8
         {
             return Some(0.0);
         }
